+++ openpathsampling/pathsimulator.py
@@ -1,10 +1,10 @@
-from openpathsampling.todict import restores_as_stub_object
-import openpathsampling as paths
-
-from openpathsampling.pathmover import PathMover
-
 import logging
+
+from openpathsampling.util.todict import restores_as_stub_object
+import openpathsampling as paths
+from openpathsampling.pathmover import PathMover
 from ops_logging import initialization_logging
+
 logger = logging.getLogger(__name__)
 init_log = logging.getLogger('openpathsampling.initialization')
