+++ openpathsampling/util/chaindict.py
